requirements.txt
streamlit>=1.35.0
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-nvidia>=0.1.0
llama-index-llms-nvidia>=0.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0
